+++ src/Hyperar.HPA.Application/Models/TeamSelection/Series.cs
@@ -2,7 +2,7 @@
 {
     public class Series
     {
-        public uint HattrickId { get; set; }
+        public long HattrickId { get; set; }
 
         public string Name { get; set; } = string.Empty;
